+++ CHANGELOG.rst
@@ -19,6 +19,13 @@ Fixed
 
 * Correct the project's dependencies: Enum34 is only required for Python versions < 3.4.
 
+* Add the :mod:`~docx_utils.exceptions` module: Exception hierarchy for the docx-utils package.
+
+* Fix `#1 <https://github.com/tantale/docx_utils/issues/1>`_:
+  Add the *on_error* option in the :func:`~docx_utils.flatten.opc_to_flat_opc` function
+  in order to ignore (or raise an exception) when a part URI cannot be resolved
+  during the Microsoft Office document parsing.
+
 Other
 ~~~~~